requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
boto==2.49.0
boto3==1.9.167
botocore==1.12.253
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.7
colorama==0.4.4
coverage==4.5.3
cryptography==35.0.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.0.5
django-cors-headers==3.11.0
django-storages==1.7.1
django-swagger-ui==0.1.11
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
docutils==0.15.2
entrypoints==0.3
enum34==1.1.6
freezegun==0.3.12
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Logentries==0.17
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mock==3.0.5
more-itertools==8.10.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nbopen==0.6
nest-asyncio==1.5.1
notebook==6.4.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py==1.10.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
python-dateutil==2.8.1
pytz==2021.3
pywin32==301
pywinpty==1.1.3
PyYAML==6.0
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.11.2
r7insight-python==0.9.2
raven==6.10.0
requests==2.26.0
requests-kerberos==0.12.0
s3transfer==0.2.1
Send2Trash==1.7.1
six==1.16.0
sqlparse==0.4.2
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
widgetsnbextension==3.5.2
winkerberos==0.8.0
zipp==3.6.0
